PiWiFi
This is to be open source of my app which extends the functionality of the WakeMeByBeacon app.  By listening for a specific UUID braodcast, this app will turn on one's wifi when they get home.  More functionality can be added for example, settnig the UUID in QSettings. Or adding a Settings page where wifi is not the only action taken, maybe send an email or text